+++ app/src/main/java/com/runnect/runnect/data/dto/request/RequestPutMyDrawDto.kt
@@ -5,7 +5,7 @@ import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class RequestPutMyDrawDto(
+data class RequestPutMyDrawDTO(
     @SerialName("courseIdList")
     val courseIdList: List<Int>
 )
