+++ app/src/main/java/com/runnect/runnect/data/dto/response/ResponsePostCourseDTO.kt
@@ -5,7 +5,7 @@ import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class ResponsePutMyDrawDto(
+data class ResponsePostCourseDTO(
     @SerialName("data")
     val data: Data,
     @SerialName("message")
@@ -13,11 +13,19 @@ data class ResponsePutMyDrawDto(
     @SerialName("status")
     val status: Int,
     @SerialName("success")
-    val success: Boolean
+    val success: Boolean,
 ) {
     @Serializable
     data class Data(
-        @SerialName("deletedCourseCount")
-        val deletedCourseCount: Int
+        @SerialName("course")
+        val course: Course,
+    )
+
+    @Serializable
+    data class Course(
+        @SerialName("createdAt")
+        val createdAt: String,
+        @SerialName("id")
+        val id: Int,
     )
 }
