+++ app/src/main/java/com/runnect/runnect/data/dto/response/ResponsePostRecordDTO.kt
@@ -5,7 +5,7 @@ import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class ResponsePostRecordDto(
+data class ResponsePostRecordDTO(
     @SerialName("data")
     val data: Data,
     @SerialName("message")
