--- app/src/main/java/com/runnect/runnect/data/dto/tmap/SearchResponseTmapDto.kt
@@ -1,10 +1,0 @@
-package com.runnect.runnect.data.dto.tmap
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class SearchResponseTmapDto(
-    @SerialName("searchPoiInfo")
-    val searchPoiInfo: SearchPoiInfo,
-)
